Struct braid::Edge

pub struct Edge {
    pub key: EdgeKey,
    pub weight: Weight,
    pub created_datetime: DateTime<UTC>,
}

An edge.

Edges are how you would represent a verb or a relationship in the datastore. An example might be "liked" or "reviewed". Edges are typed and directed.

Fields
key: EdgeKey

The key to the edge.

weight: Weight

The weight of the edge.

created_datetime: DateTime<UTC>

When the edge was created.

Methods

impl Edge

fn new_with_current_datetime(key: EdgeKey, weight: Weight) -> Edge

Creates a new edge with the current datetime in UTC.

Arguments

key - The key to the edge.

weight - The edge weight.

inbound_id - The id of the inbound vertex.

fn new(key: EdgeKey, weight: Weight, created_datetime: DateTime<UTC>) -> Edge

Creates a new edge with a specified datetime.

Arguments

key - The key to the edge.

weight - The weight of the edge.

created_datetime - When the edge was created.