Struct braid::PostgresDatastore

pub struct PostgresDatastore { /* fields omitted */ }

A datastore that is backed by a postgres database.

Methods

impl PostgresDatastore

fn new(pool_size: Option<u32>,
connection_string: String,
secret: String,
secure_uuids: bool)
-> PostgresDatastore

Creates a new postgres-backed datastore.

Arguments

pool_size - The maximum number of connections to maintain to postgres. If None, it defaults to twice the number of CPUs.

connetion_string - The postgres database connection string.

secret - A secret value. This is used as a pepper in hashing sensitive account data.

secure_uuids - If true, UUIDv4 will be used, which will result in difficult to guess UUIDs at the detriment of a more index-optimized (and thus faster) variant.

fn create_schema(connection_string: String) -> Result<(), Error>

Creates a new postgres-backed datastore.

Arguments

connetion_string - The postgres database connection string.