Module braid::tests

Unit tests for datastore implementations.

These are exported so that datastore implementations outside of the braid crate can reuse them. Generally you can use the convenience macros test_account_impl, test_metadata_impl, and test_transaction_impl rather than referencing these functions directly - but they may be useful to reference directly if you want to implement a datastore that doesn't support full functionality.

Structs

DatastoreTestSandbox

Functions

create_edge_from

create_edges

create_time_range_queryable_edges

should_create_a_valid_edge

should_delete_a_valid_edge

should_delete_a_valid_vertex

should_fail_auth_with_a_bad_password

should_fail_auth_with_a_bad_username

should_fail_to_lookup_invalid_accounts

should_fail_when_attempting_to_delete_invalid_accounts

should_get_a_valid_edge

should_get_all_vertices

should_get_all_vertices_out_of_range

should_get_all_vertices_with_zero_limit

should_get_an_edge_count

should_get_an_edge_count_for_an_invalid_edge

should_get_an_edge_count_with_no_type

should_get_an_edge_range

should_get_edges

should_get_edges_with_no_high

should_get_edges_with_no_low

should_get_edges_with_no_time

should_get_edges_with_no_type

should_get_no_edges_for_an_invalid_range

should_get_no_edges_for_reversed_time

should_get_single_vertices

should_get_single_vertices_nonexisting

should_get_vertices

should_get_vertices_piped

should_handle_account_metadata

should_handle_edge_metadata

should_handle_global_metadata

should_handle_vertex_metadata

should_lookup_valid_accounts

should_not_create_an_edge_with_bad_permissions

should_not_create_an_invalid_edge

should_not_delete_an_edge_with_bad_permissions

should_not_delete_an_invalid_edge

should_not_delete_an_invalid_vertex

should_not_delete_an_unowned_vertex

should_not_delete_invalid_account_metadata

should_not_delete_invalid_edge_metadata

should_not_delete_invalid_vertex_metadata

should_not_get_an_invalid_edge

should_not_set_invalid_account_metadata

should_not_set_invalid_edge_metadata

should_not_set_invalid_vertex_metadata

should_successfully_auth_with_good_credentials