Trait braid::Datastore

pub trait Datastore<T: Transaction> {
    fn has_account(&self, account_id: Uuid) -> Result<bool, Error>;
    fn create_account(&self) -> Result<(Uuid, String), Error>;
    fn delete_account(&self, account_id: Uuid) -> Result<(), Error>;
    fn auth(&self, account_id: Uuid, secret: String) -> Result<bool, Error>;
    fn transaction(&self, account_id: Uuid) -> Result<T, Error>;
}

Specifies a datastore implementation.

Datastores are responsible for managing accounts, and constructing new transactions.

Errors

All methods may return an error if something unexpected happens - e.g. if there was a problem connecting to the underlying database.

Required Methods

fn has_account(&self, account_id: Uuid) -> Result<bool, Error>

Checks if an account exists.

Arguments

account_id - The ID of the account to check.

fn create_account(&self) -> Result<(Uuid, String), Error>

Creates a new account, returning a tuple of its ID and secret.

fn delete_account(&self, account_id: Uuid) -> Result<(), Error>

Deletes an account.

Arguments

account_id - The ID of the account to delete.

Errors

Returns an error if the account does not exist.

fn auth(&self, account_id: Uuid, secret: String) -> Result<bool, Error>

Checks account authentication.

Arguments

account_id - The ID of the account to authenticate.

secret - The account's secret.

fn transaction(&self, account_id: Uuid) -> Result<T, Error>

Creates a new transaction tied to a given account.

Arguments

account_id - The ID of the account that's triggering the transaction.