Trait braid::Transaction

pub trait Transaction {
    fn create_vertex(&self, t: Type) -> Result<Uuid, Error>;
    fn get_vertices(&self, q: VertexQuery) -> Result<Vec<Vertex>, Error>;
    fn delete_vertices(&self, q: VertexQuery) -> Result<(), Error>;
    fn create_edge(&self, key: EdgeKey, weight: Weight) -> Result<(), Error>;
    fn get_edges(&self, q: EdgeQuery) -> Result<Vec<Edge>, Error>;
    fn delete_edges(&self, q: EdgeQuery) -> Result<(), Error>;
    fn get_edge_count(&self, q: EdgeQuery) -> Result<u64, Error>;
    fn get_global_metadata(&self, name: String) -> Result<JsonValue, Error>;
    fn set_global_metadata(&self,
                           name: String,
                           value: JsonValue)
                           -> Result<(), Error>;
    fn delete_global_metadata(&self, name: String) -> Result<(), Error>;
    fn get_account_metadata(&self,
                            account_id: Uuid,
                            name: String)
                            -> Result<JsonValue, Error>;
    fn set_account_metadata(&self,
                            account_id: Uuid,
                            name: String,
                            value: JsonValue)
                            -> Result<(), Error>;
    fn delete_account_metadata(&self,
                               account_id: Uuid,
                               name: String)
                               -> Result<(), Error>;
    fn get_vertex_metadata(&self,
                           q: VertexQuery,
                           name: String)
                           -> Result<HashMap<Uuid, JsonValue>, Error>;
    fn set_vertex_metadata(&self,
                           q: VertexQuery,
                           name: String,
                           value: JsonValue)
                           -> Result<(), Error>;
    fn delete_vertex_metadata(&self,
                              q: VertexQuery,
                              name: String)
                              -> Result<(), Error>;
    fn get_edge_metadata(&self,
                         q: EdgeQuery,
                         name: String)
                         -> Result<HashMap<EdgeKey, JsonValue>, Error>;
    fn set_edge_metadata(&self,
                         q: EdgeQuery,
                         name: String,
                         value: JsonValue)
                         -> Result<(), Error>;
    fn delete_edge_metadata(&self,
                            q: EdgeQuery,
                            name: String)
                            -> Result<(), Error>;
    fn commit(self) -> Result<(), Error>;
    fn rollback(self) -> Result<(), Error>;
}

Specifies a transaction implementation, which are returned by datastores. Transactions are responsible for managing:

Vertices.

Edges, which connect two vertices.

Global metadata: metadata that is not owned by anything, and as aresult needs to be manually managed.

Account metadata: metadata that is owned by an account, and will beautomatically deleted when the associated account is deleted.

Vertex metadata: metadata that is owned by a vertex, and will beautomatically deleted when the associated vertex is deleted.

Edge metadata: metadata that is owned by an edge, and will beautomatically deleted when the associated edge is deleted.

Required Methods

fn create_vertex(&self, t: Type) -> Result<Uuid, Error>

Creates a new vertex.

Arguments

t - The type of the vertex.

fn get_vertices(&self, q: VertexQuery) -> Result<Vec<Vertex>, Error>

Gets a range of vertices specified by a query.

Arguments

q - The query to run.

fn delete_vertices(&self, q: VertexQuery) -> Result<(), Error>

Deletes existing vertices specified by a query.

Arguments

q - The query to run.

fn create_edge(&self, key: EdgeKey, weight: Weight) -> Result<(), Error>

Creates a new edge. If the edge already exists, this will update it with a new update datetime and weight. The transaction tied to the account must own the vertex from which the edge is outbounding from, but does not need to own the inbounding vertex.

Arguments

key: The edge to create.

weight - The edge weight.

Errors

Return Error::VertexNotFound if either of the connected vertices do not exist. Returns Error::Unauthorized if the account tied to the current transaction does not own the source vertex.

fn get_edges(&self, q: EdgeQuery) -> Result<Vec<Edge>, Error>

Gets a range of edges specified by a query.

Arguments

q - The query to run.

fn delete_edges(&self, q: EdgeQuery) -> Result<(), Error>

Deletes a set of edges specified by a query.

Arguments

q - The query to run.

fn get_edge_count(&self, q: EdgeQuery) -> Result<u64, Error>

Gets the number of edges that match a query.

Arguments

outbound_id - The ID of the outbound vertex.

t - The edge type.

fn get_global_metadata(&self, name: String) -> Result<JsonValue, Error>

Gets a global metadata value.

Arguments

name - The metadata name.

Errors

Returns Error::MetadataNotFound if the metadata does not exist.

fn set_global_metadata(&self,
name: String,
value: JsonValue)
-> Result<(), Error>

Sets a global metadata value.

Arguments

name - The metadata name.

value - The metadata value.

fn delete_global_metadata(&self, name: String) -> Result<(), Error>

Deletes a global metadata value.

Arguments

name - The metadata name.

Errors

Returns Error::MetadataNotFound if the metadata does not exist.

fn get_account_metadata(&self,
account_id: Uuid,
name: String)
-> Result<JsonValue, Error>

Gets an account metadata value.

Arguments

account_id: The ID of the account that the metadata is tied to.

name - The metadata name.

Errors

Returns Error::MetadataNotFound if the metadata does not exist.

fn set_account_metadata(&self,
account_id: Uuid,
name: String,
value: JsonValue)
-> Result<(), Error>

Sets an account metadata value.

Arguments

account_id: The ID of the account that the metadata is tied to.

name - The metadata name.

value - The metadata value.

Errors

Returns Error::AccountNotFound if the specified account ID does not exist.

fn delete_account_metadata(&self,
account_id: Uuid,
name: String)
-> Result<(), Error>

Deletes an account metadata value.

Arguments

account_id: The ID of the account that the metadata is tied to.

name - The metadata name.

Errors

Returns Error::MetadataNotFound if the metadata does not exist.

fn get_vertex_metadata(&self,
q: VertexQuery,
name: String)
-> Result<HashMap<Uuid, JsonValue>, Error>

Gets a vertex metadata value.

Arguments

q - The query to run.

name - The metadata name.

fn set_vertex_metadata(&self,
q: VertexQuery,
name: String,
value: JsonValue)
-> Result<(), Error>

Sets a vertex metadata value.

Arguments

q - The query to run.

name - The metadata name.

value - The metadata value.

fn delete_vertex_metadata(&self,
q: VertexQuery,
name: String)
-> Result<(), Error>

Deletes a vertex metadata value.

Arguments

q - The query to run.

name - The metadata name.

fn get_edge_metadata(&self,
q: EdgeQuery,
name: String)
-> Result<HashMap<EdgeKey, JsonValue>, Error>

Gets an edge metadata value.

Arguments

q - The query to run.

name - The metadata name.

fn set_edge_metadata(&self,
q: EdgeQuery,
name: String,
value: JsonValue)
-> Result<(), Error>

Sets an edge metadata value.

Arguments

q - The query to run.

name - The metadata name.

value - The metadata value.

fn delete_edge_metadata(&self, q: EdgeQuery, name: String) -> Result<(), Error>

Deletes an edge metadata value.

Arguments

q - The query to run.

name - The metadata name.

fn commit(self) -> Result<(), Error>

Commits the transaction.

fn rollback(self) -> Result<(), Error>

Rolls the transaction back.