Crate braid

Braid - a graph datastore.

Braid is broken up into a library and an application. This is the library, which you would use if you want to create new datastore implementations, or plug into the low-level details of braid. For most use cases, you can use the application, which exposes an API and scripting layer.

Modules

tests
Unit tests for datastore implementations.

Macros

define_test
Defines a unit test function.

test_account_impl
Use this macro to enable the standard test suite for accounts.

test_metadata_impl
Use this macro to enable the standard test suite for metadata.

test_transaction_impl
Use this macro to enable the standard test suite for transactions.

Structs

Edge
An edge.

EdgeKey
Represents a uniquely identifiable key to an edge.

PostgresDatastore
A datastore that is backed by a postgres database.

PostgresTransaction
A postgres-backed datastore transaction.

RocksdbDatastore
A datastore that is backed by rocksdb.

RocksdbTransaction
A transaction that is backed by rocksdb.

Type
An edge or vertex type.

ValidationError
The error returned when there is an attempt to instantiate a model with an invalid value.

Vertex
A vertex.

Weight
An edge weight.

Enums

EdgeQuery
A query for edges.

Error
The error returned by datastore and transaction implementation methods.

QueryTypeConverter
Specifies what kind of items should be piped from one type of query to another.

VertexQuery
A query for vertices.

Traits

Datastore
Specifies a datastore implementation.

Transaction
Specifies a transaction implementation, which are returned by datastores. Transactions are responsible for managing: